requirements.txt
aiosqlite==0.21.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
Automat==25.4.16
beautifulsoup4==4.13.5
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
constantly==23.10.4
decorator==5.2.1
exceptiongroup==1.2.2
fastapi==0.116.1
google-auth==2.40.3
google-genai==1.32.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
nodeenv==1.9.1
objgraph==3.6.2
outcome==1.3.0.post0
packaging==25.0
pluggy==1.6.0
prisma==0.15.0
psutil==7.0.0
psycopg==3.2.9
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-tornasync==0.6.0.post2
pytest-trio==0.8.0
pytest-twisted==1.14.3
python-dotenv==1.0.1
python-multipart==0.0.20
pywin32==311
PyYAML==6.0.2
requests==2.32.5
rsa==4.9.1
setuptools==80.9.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.38
starlette==0.45.3
StrEnum==0.4.15
tenacity==9.1.2
tomlkit==0.13.2
tornado==6.5.2
trio==0.30.0
Twisted==25.5.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.0
watchfiles==1.1.0
websockets==15.0.1
zope.interface==7.2
